♻️  Move withCookieLockAccess to SessionStore and rename processStorageOperations
