+++ packages/core/src/domain/session/sessionCookieStore.ts
@@ -1,11 +1,9 @@
 import type { CookieOptions } from '../../browser/cookie'
 import { deleteCookie, getCookie, setCookie } from '../../browser/cookie'
-import { setTimeout } from '../../tools/timer'
 import { isChromium } from '../../tools/utils/browserDetection'
 import { dateNow } from '../../tools/utils/timeUtils'
 import { objectEntries } from '../../tools/utils/polyfills'
 import { isEmptyObject } from '../../tools/utils/objectUtils'
-import { generateUUID } from '../../tools/utils/stringUtils'
 import { SESSION_EXPIRATION_DELAY } from './sessionConstants'
 import type { SessionState } from './sessionStorage'
 
@@ -13,105 +11,6 @@ const SESSION_ENTRY_REGEXP = /^([a-z]+)=([a-z0-9-]+)$/
 const SESSION_ENTRY_SEPARATOR = '&'
 
 export const SESSION_COOKIE_NAME = '_dd_s'
-
-// arbitrary values
-export const LOCK_RETRY_DELAY = 10
-export const MAX_NUMBER_OF_LOCK_RETRIES = 100
-
-type Operations = {
-  options: CookieOptions
-  process: (cookieSession: SessionState) => SessionState | undefined
-  after?: (cookieSession: SessionState) => void
-}
-
-const bufferedOperations: Operations[] = []
-let ongoingOperations: Operations | undefined
-
-export function withCookieLockAccess(operations: Operations, numberOfRetries = 0) {
-  if (!ongoingOperations) {
-    ongoingOperations = operations
-  }
-  if (operations !== ongoingOperations) {
-    bufferedOperations.push(operations)
-    return
-  }
-  if (numberOfRetries >= MAX_NUMBER_OF_LOCK_RETRIES) {
-    next()
-    return
-  }
-  let currentLock: string
-  let currentSession = retrieveSessionCookie()
-  if (isCookieLockEnabled()) {
-    // if someone has lock, retry later
-    if (currentSession.lock) {
-      retryLater(operations, numberOfRetries)
-      return
-    }
-    // acquire lock
-    currentLock = generateUUID()
-    currentSession.lock = currentLock
-    setSessionCookie(currentSession, operations.options)
-    // if lock is not acquired, retry later
-    currentSession = retrieveSessionCookie()
-    if (currentSession.lock !== currentLock) {
-      retryLater(operations, numberOfRetries)
-      return
-    }
-  }
-  let processedSession = operations.process(currentSession)
-  if (isCookieLockEnabled()) {
-    // if lock corrupted after process, retry later
-    currentSession = retrieveSessionCookie()
-    if (currentSession.lock !== currentLock!) {
-      retryLater(operations, numberOfRetries)
-      return
-    }
-  }
-  if (processedSession) {
-    persistSessionCookie(processedSession, operations.options)
-  }
-  if (isCookieLockEnabled()) {
-    // correctly handle lock around expiration would require to handle this case properly at several levels
-    // since we don't have evidence of lock issues around expiration, let's just not do the corruption check for it
-    if (!(processedSession && isExpiredState(processedSession))) {
-      // if lock corrupted after persist, retry later
-      currentSession = retrieveSessionCookie()
-      if (currentSession.lock !== currentLock!) {
-        retryLater(operations, numberOfRetries)
-        return
-      }
-      delete currentSession.lock
-      setSessionCookie(currentSession, operations.options)
-      processedSession = currentSession
-    }
-  }
-  // call after even if session is not persisted in order to perform operations on
-  // up-to-date cookie value, the value could have been modified by another tab
-  operations.after?.(processedSession || currentSession)
-  next()
-}
-
-/**
- * Cookie lock strategy allows mitigating issues due to concurrent access to cookie.
- * This issue concerns only chromium browsers and enabling this on firefox increase cookie write failures.
- */
-function isCookieLockEnabled() {
-  return isChromium()
-}
-
-function retryLater(operations: Operations, currentNumberOfRetries: number) {
-  setTimeout(() => {
-    withCookieLockAccess(operations, currentNumberOfRetries + 1)
-  }, LOCK_RETRY_DELAY)
-}
-
-function next() {
-  ongoingOperations = undefined
-  const nextOperations = bufferedOperations.shift()
-  if (nextOperations) {
-    withCookieLockAccess(nextOperations)
-  }
-}
 
 export function persistSessionCookie(session: SessionState, options: CookieOptions) {
   if (isExpiredState(session)) {
@@ -122,7 +21,7 @@ export function persistSessionCookie(session: SessionState, options: CookieOptio
   setSessionCookie(session, options)
 }
 
-function setSessionCookie(session: SessionState, options: CookieOptions) {
+export function setSessionCookie(session: SessionState, options: CookieOptions) {
   setCookie(SESSION_COOKIE_NAME, toSessionString(session), SESSION_EXPIRATION_DELAY, options)
 }
 
@@ -158,6 +57,14 @@ function isValidSessionString(sessionString: string | undefined): sessionString 
   )
 }
 
-function isExpiredState(session: SessionState) {
+export function isExpiredState(session: SessionState) {
   return isEmptyObject(session)
+}
+
+/**
+ * Cookie lock strategy allows mitigating issues due to concurrent access to cookie.
+ * This issue concerns only chromium browsers and enabling this on firefox increase cookie write failures.
+ */
+export function isCookieLockEnabled() {
+  return isChromium()
 }
