+++ packages/core/src/domain/session/sessionStore.ts
@@ -1,12 +1,19 @@
 import type { CookieOptions } from '../../browser/cookie'
 import { COOKIE_ACCESS_DELAY } from '../../browser/cookie'
-import { clearInterval, setInterval } from '../../tools/timer'
+import { clearInterval, setInterval, setTimeout } from '../../tools/timer'
 import { Observable } from '../../tools/observable'
 import { dateNow } from '../../tools/utils/timeUtils'
 import { throttle } from '../../tools/utils/functionUtils'
 import { generateUUID } from '../../tools/utils/stringUtils'
 import { SESSION_TIME_OUT_DELAY } from './sessionConstants'
-import { deleteSessionCookie, retrieveSessionCookie, withCookieLockAccess } from './sessionCookieStore'
+import {
+  deleteSessionCookie,
+  isCookieLockEnabled,
+  isExpiredState,
+  persistSessionCookie,
+  retrieveSessionCookie,
+  setSessionCookie,
+} from './sessionCookieStore'
 import type { SessionState } from './sessionStorage'
 
 export interface SessionStore {
@@ -38,7 +45,7 @@ export function startSessionStore<TrackingType extends string>(
 
   function expandOrRenewSession() {
     let isTracked: boolean
-    withCookieLockAccess({
+    processStorageOperations({
       options,
       process: (sessionState) => {
         const synchronizedSession = synchronizeSession(sessionState)
@@ -55,7 +62,7 @@ export function startSessionStore<TrackingType extends string>(
   }
 
   function expandSession() {
-    withCookieLockAccess({
+    processStorageOperations({
       options,
       process: (sessionState) => (hasSessionInCache() ? synchronizeSession(sessionState) : undefined),
     })
@@ -67,7 +74,7 @@ export function startSessionStore<TrackingType extends string>(
    * - if the session is not active, clear the session cookie and expire the session cache
    */
   function watchSession() {
-    withCookieLockAccess({
+    processStorageOperations({
       options,
       process: (sessionState) => (!isActiveSession(sessionState) ? {} : undefined),
       after: synchronizeSession,
@@ -146,5 +153,96 @@ export function startSessionStore<TrackingType extends string>(
     stop: () => {
       clearInterval(watchSessionTimeoutId)
     },
+  }
+}
+
+// arbitrary values
+export const LOCK_RETRY_DELAY = 10
+export const MAX_NUMBER_OF_LOCK_RETRIES = 100
+
+type Operations = {
+  options: CookieOptions
+  process: (cookieSession: SessionState) => SessionState | undefined
+  after?: (cookieSession: SessionState) => void
+}
+
+const bufferedOperations: Operations[] = []
+let ongoingOperations: Operations | undefined
+
+export function processStorageOperations(operations: Operations, numberOfRetries = 0) {
+  if (!ongoingOperations) {
+    ongoingOperations = operations
+  }
+  if (operations !== ongoingOperations) {
+    bufferedOperations.push(operations)
+    return
+  }
+  if (numberOfRetries >= MAX_NUMBER_OF_LOCK_RETRIES) {
+    next()
+    return
+  }
+  let currentLock: string
+  let currentSession = retrieveSessionCookie()
+  if (isCookieLockEnabled()) {
+    // if someone has lock, retry later
+    if (currentSession.lock) {
+      retryLater(operations, numberOfRetries)
+      return
+    }
+    // acquire lock
+    currentLock = generateUUID()
+    currentSession.lock = currentLock
+    setSessionCookie(currentSession, operations.options)
+    // if lock is not acquired, retry later
+    currentSession = retrieveSessionCookie()
+    if (currentSession.lock !== currentLock) {
+      retryLater(operations, numberOfRetries)
+      return
+    }
+  }
+  let processedSession = operations.process(currentSession)
+  if (isCookieLockEnabled()) {
+    // if lock corrupted after process, retry later
+    currentSession = retrieveSessionCookie()
+    if (currentSession.lock !== currentLock!) {
+      retryLater(operations, numberOfRetries)
+      return
+    }
+  }
+  if (processedSession) {
+    persistSessionCookie(processedSession, operations.options)
+  }
+  if (isCookieLockEnabled()) {
+    // correctly handle lock around expiration would require to handle this case properly at several levels
+    // since we don't have evidence of lock issues around expiration, let's just not do the corruption check for it
+    if (!(processedSession && isExpiredState(processedSession))) {
+      // if lock corrupted after persist, retry later
+      currentSession = retrieveSessionCookie()
+      if (currentSession.lock !== currentLock!) {
+        retryLater(operations, numberOfRetries)
+        return
+      }
+      delete currentSession.lock
+      setSessionCookie(currentSession, operations.options)
+      processedSession = currentSession
+    }
+  }
+  // call after even if session is not persisted in order to perform operations on
+  // up-to-date cookie value, the value could have been modified by another tab
+  operations.after?.(processedSession || currentSession)
+  next()
+}
+
+function retryLater(operations: Operations, currentNumberOfRetries: number) {
+  setTimeout(() => {
+    processStorageOperations(operations, currentNumberOfRetries + 1)
+  }, LOCK_RETRY_DELAY)
+}
+
+function next() {
+  ongoingOperations = undefined
+  const nextOperations = bufferedOperations.shift()
+  if (nextOperations) {
+    processStorageOperations(nextOperations)
   }
 }
